moving hashtags to their own module
