+++ core/ui/hashtagcard/build.gradle.kts
@@ -1,0 +1,25 @@
+plugins {
+    id("social.firefly.android.library")
+    id("social.firefly.android.library.compose")
+}
+
+android {
+    namespace = "social.firefly.core.ui.hashtagcard"
+}
+
+dependencies {
+    implementation(project(":core:model"))
+    implementation(project(":core:designsystem"))
+    implementation(project(":core:ui:common"))
+    implementation(project(":core:common"))
+    implementation(project(":core:repository:mastodon"))
+    implementation(project(":core:navigation"))
+
+    implementation(libs.androidx.paging.compose)
+
+    implementation(libs.androidx.navigation.compose)
+    implementation(libs.koin.core)
+    implementation(libs.koin.androidx.compose)
+
+    implementation(libs.jakewharton.timber)
+}
