+++ core/ui/hashtagcard/src/main/java/social/firefly/core/ui/hashtagcard/HashtagInteractions.kt
@@ -1,0 +1,8 @@
+package social.firefly.core.ui.hashtagcard
+
+import social.firefly.core.ui.common.following.FollowStatus
+
+interface HashtagInteractions {
+    fun onHashTagFollowClicked(name: String, followStatus: FollowStatus)
+    fun onHashtagClick(name: String)
+}
