+++ feature/discover/build.gradle.kts
@@ -13,6 +13,7 @@ dependencies {
     implementation(project(":core:designsystem"))
     implementation(project(":core:ui:common"))
     implementation(project(":core:ui:postcard"))
+    implementation(project(":core:ui:hashtagcard"))
     implementation(project(":core:ui:accountfollower"))
     implementation(project(":core:common"))
     implementation(project(":core:navigation"))
