+++ settings.gradle.kts
@@ -71,3 +71,4 @@ include(":feature:notifications")
 include(":core:ui:notifications")
 include(":feature:media")
 include(":core:push")
+include(":core:ui:hashtagcard")
